added frontend

+++ backend/controllers/product.controller.js
@@ -48,13 +48,17 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
         const {id} = req.params;
+        if (mongoose.Types.ObjectId.isValid(id) === false) {
+            return res.status(400).json({ message: "Invalid product ID" });
     
+        
+        }
         try {
              await Product.findByIdAndDelete(id);
              res.status(200).json({ success: true, message: "Product deleted successfully" });
             
         } catch (error) {
             console.error("Error in deleting product: ",error.message);
-            res.status(500).json({ success: false, message: "Product not found" });
+            res.status(500).json({ success: false, message: "Internal Server Error" });
         }
 }
